% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_history.R
\name{get_history}
\alias{get_history}
\title{get_history}
\usage{
get_history(user)
}
\arguments{
\item{user}{Last.fm username}
}
\value{
A lastfm scrobble, or music listening history dataframe
}
\description{
This function pulls a last.fm user's music listening history
}
