% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_summaries.R
\name{get_session_df}
\alias{get_session_df}
\title{get_session_summaries}
\usage{
get_session_df(audio_features)
}
\arguments{
\item{audio_features}{is a last.fm scrobble/spotify audio features dataframe, as created by get_history_audio_features.}
}
\value{
a spotify/last.fm dataframe with computed values on date, sessions, and listening times
}
\description{
This function performs further variable computation to aid in analysis, including date manipulation, track start/end times, differentials to help determine skipping, as well as session length, determined by minute long difference in music listening end and start times.
}
